+++ modules/framework/galasa-parent/dev.galasa.framework/src/testFixtures/java/dev/galasa/framework/mocks/MockRBACService.java
@@ -122,7 +122,7 @@ public class MockRBACService implements RBACService {
     }
 
     @Override
-    public long getUserPriority(String loginId) throws RBACException {
+    public int getUserPriority(String loginId) throws RBACException {
         return usersCache.getUserPriority(loginId);
     }
 }
